Fixed namespaces and moved SoundManager
- Fixed scripts that lacked namespaces (SliderSelect, SoundManager)
- Moved Sound Manager to Audio folder
